Refactoring location of all the util codes
