+++ src/main/java/seedu/todo/ui/view/CommandErrorView.java
@@ -8,10 +8,11 @@ import javafx.scene.layout.GridPane;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 import seedu.todo.commons.core.LogsCenter;
-import seedu.todo.commons.util.FxViewUtil;
+import seedu.todo.ui.util.FxViewUtil;
 import seedu.todo.model.ErrorBag;
 import seedu.todo.ui.UiPart;
 import seedu.todo.ui.UiPartLoader;
+import seedu.todo.ui.util.ViewGeneratorUtil;
 
 import java.util.List;
 import java.util.Map;
@@ -138,8 +139,8 @@ public class CommandErrorView extends UiPart {
      * @param rightText text for the second column
      */
     private void addRowToGrid(GridPane targetGrid, int rowIndex, String leftText, String rightText) {
-        Label leftLabel = FxViewUtil.constructLabel(leftText, STYLE_TEXT_4);
-        Label rightLabel = FxViewUtil.constructLabel(rightText, STYLE_TEXT_4);
+        Label leftLabel = ViewGeneratorUtil.constructLabel(leftText, STYLE_TEXT_4);
+        Label rightLabel = ViewGeneratorUtil.constructLabel(rightText, STYLE_TEXT_4);
         targetGrid.addRow(rowIndex, leftLabel, rightLabel);
     }
