+++ src/main/java/seedu/todo/ui/view/CommandFeedbackView.java
@@ -6,9 +6,10 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 import seedu.todo.commons.core.LogsCenter;
-import seedu.todo.commons.util.FxViewUtil;
+import seedu.todo.ui.util.FxViewUtil;
 import seedu.todo.ui.UiPart;
 import seedu.todo.ui.UiPartLoader;
+import seedu.todo.ui.util.ViewStyleUtil;
 
 import java.util.logging.Logger;
 
@@ -67,14 +68,14 @@ public class CommandFeedbackView extends UiPart {
      * Indicate an error visually on the {@link #commandFeedbackLabel}
      */
     public void flagError() {
-        FxViewUtil.addClassStyle(commandFeedbackLabel, ERROR_STYLE);
+        ViewStyleUtil.addClassStyle(commandFeedbackLabel, ERROR_STYLE);
     }
 
     /**
      * Remove the error flag visually on the {@link #commandFeedbackLabel}
      */
     public void unFlagError() {
-        FxViewUtil.removeClassStyle(commandFeedbackLabel, ERROR_STYLE);
+        ViewStyleUtil.removeClassStyle(commandFeedbackLabel, ERROR_STYLE);
     }
 
     /* Override Methods */
